refactor(validation): 특정 문자만 포함되어 있는 지 정규식으로 검증하는 작업을 별도의 추상 클래스로 분리
